Range.jsx and IntegralResizable.jsx now use Resizable.jsx
